Migration to SDL3
1) Migrated to SDL 3.2.26.
2) Audio temporaly removed. Will be added as soon as it is released.
3) Add modal mode for a window. Now can set owner. MainWindow sets via raw pointer and app gets ownership.
4) Added ThemeMode.
5) Updated dependecies list in README.

+++ Sgl/Input/TextEventArgs.h
@@ -7,8 +7,8 @@ namespace Sgl
 	struct TextEditingEventArgs
 	{
 		std::string Text;
-		size_t SelectionStart;
-		size_t SelectionLength;
+		int SelectionStart;
+		int SelectionLength;
 	};
 
 	struct TextInputEventArgs
